# swagger-spec.json
{"swagger":"2.0","info":{"title":"Arbi Service API","description":"Arbi Service API documentation","version":"1.0.0"},"definitions":{},"paths":{"/api/bots":{"post":{"summary":"Creates a new bot with the given name and type","description":"Create a new bot","tags":["Bots"],"parameters":[{"name":"body","in":"body","schema":{"type":"object","required":["name","type"],"properties":{"name":{"type":"string","minLength":1,"maxLength":255,"description":"Name of the bot"},"type":{"type":"string","minLength":1,"maxLength":255,"description":"Type of the bot"}},"additionalProperties":false}}],"responses":{"201":{"description":"Default Response","schema":{"type":"object","properties":{"id":{"type":"string","format":"uuid","description":"Unique identifier of the bot"},"name":{"type":"string","description":"Name of the bot"},"type":{"type":"string","description":"Type of the bot"},"status":{"type":"string","enum":["active","stopped","paused"],"description":"Current status of the bot"},"config":{"type":"object","description":"Configuration settings for the bot"},"created_at":{"type":"string","format":"date-time","description":"Timestamp when the bot was created"},"updated_at":{"type":"string","format":"date-time","description":"Timestamp when the bot was last updated"}}}},"400":{"description":"Default Response","schema":{"type":"object","properties":{"message":{"type":"string","description":"Error message"}}}}}},"get":{"summary":"Retrieves all bots in the system","description":"Get all bots","tags":["Bots"],"responses":{"200":{"description":"Default Response","schema":{"type":"array","items":{"type":"object","properties":{"id":{"type":"string","format":"uuid","description":"Unique identifier of the bot"},"name":{"type":"string","description":"Name of the bot"},"type":{"type":"string","description":"Type of the bot"},"status":{"type":"string","enum":["active","stopped","paused"],"description":"Current status of the bot"},"config":{"type":"object","description":"Configuration settings for the bot"},"created_at":{"type":"string","format":"date-time","description":"Timestamp when the bot was created"},"updated_at":{"type":"string","format":"date-time","description":"Timestamp when the bot was last updated"}}}}}}}},"/api/bots/{id}":{"get":{"summary":"Retrieves a single bot by its ID","description":"Get a bot by ID","tags":["Bots"],"parameters":[{"type":"string","description":"Bot ID","format":"uuid","required":true,"in":"path","name":"id"}],"responses":{"200":{"description":"Default Response","schema":{"type":"object","properties":{"id":{"type":"string","format":"uuid","description":"Unique identifier of the bot"},"name":{"type":"string","description":"Name of the bot"},"type":{"type":"string","description":"Type of the bot"},"status":{"type":"string","enum":["active","stopped","paused"],"description":"Current status of the bot"},"config":{"type":"object","description":"Configuration settings for the bot"},"created_at":{"type":"string","format":"date-time","description":"Timestamp when the bot was created"},"updated_at":{"type":"string","format":"date-time","description":"Timestamp when the bot was last updated"}}}},"404":{"description":"Default Response","schema":{"type":"object","properties":{"message":{"type":"string","description":"Error message"}}}}}},"put":{"summary":"Updates an existing bot by ID","description":"Update an existing bot","tags":["Bots"],"parameters":[{"name":"body","in":"body","schema":{"type":"object","properties":{"name":{"type":"string","minLength":1,"maxLength":255,"description":"Name of the bot"},"type":{"type":"string","minLength":1,"maxLength":255,"description":"Type of the bot"}},"additionalProperties":false}},{"type":"string","description":"Bot ID","format":"uuid","required":true,"in":"path","name":"id"}],"responses":{"200":{"description":"Default Response","schema":{"type":"object","properties":{"id":{"type":"string","format":"uuid","description":"Unique identifier of the bot"},"name":{"type":"string","description":"Name of the bot"},"type":{"type":"string","description":"Type of the bot"},"status":{"type":"string","enum":["active","stopped","paused"],"description":"Current status of the bot"},"config":{"type":"object","description":"Configuration settings for the bot"},"created_at":{"type":"string","format":"date-time","description":"Timestamp when the bot was created"},"updated_at":{"type":"string","format":"date-time","description":"Timestamp when the bot was last updated"}}}},"404":{"description":"Default Response","schema":{"type":"object","properties":{"message":{"type":"string","description":"Error message"}}}}}},"delete":{"summary":"Deletes a bot by ID","description":"Delete a bot","tags":["Bots"],"parameters":[{"type":"string","description":"Bot ID","format":"uuid","required":true,"in":"path","name":"id"}],"responses":{"204":{"description":"Bot deleted successfully"},"404":{"description":"Default Response","schema":{"type":"object","properties":{"message":{"type":"string","description":"Error message"}}}}}}}},"host":"localhost:8080","schemes":["http"],"consumes":["application/json"],"produces":["application/json"],"tags":[{"name":"Bots","description":"Bot management endpoints"}],"externalDocs":{"url":"https://swagger.io","description":"Find more info here"}}